+++ UMI3D-SDK/Assets/UMI3D SDK/ClientDevelopmentKit/Core/Runtime/Notification/NotificationLoader.cs
@@ -62,7 +62,7 @@ namespace umi3d.cdk
                 case UMI3DPropertyKeys.NotificationObjectId:
                     var Odto = dto as NotificationOnObjectDto;
                     if (Odto == null) return false;
-                    Odto.objectId = (ulong)property.value;
+                    Odto.objectId = (ulong)(long)property.value;
                     break;
                 default:
                     return false;
